+++ jono/Cargo.toml
@@ -9,14 +9,16 @@ repository.workspace = true
 version.workspace = true
 
 [features]
-default = ["produce", "consume"]
+default = ["produce", "consume", "harvest"]
 produce = ["dep:jono_produce"]
 consume = ["dep:jono_consume"]
+harvest = ["dep:jono_harvest"]
 
 [dependencies]
 jono_core = { path = "../jono_core", version = "=0.1.6-rc.4" }
 jono_produce = { path = "../jono_produce", version = "=0.1.6-rc.4", default-features = false, optional = true }
 jono_consume = { path = "../jono_consume", version = "=0.1.6-rc.4", default-features = false, optional = true }
+jono_harvest = { path = "../jono_harvest", version = "=0.1.6-rc.4", default-features = false, optional = true }
 
 [dev-dependencies]
 serde.workspace = true
