+++ jono_harvest/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "jono_harvest"
+description = "Post-process interface for completed jobs to Jono queues"
+readme = "README.md"
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+version.workspace = true
+
+[dependencies]
+jono_core = { path = "../jono_core", version = "=0.1.6-rc.4" }
+redis.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+tracing.workspace = true
