+++ jono_harvest/src/lib.rs
@@ -1,0 +1,9 @@
+//! `jono_harvest` provides the interface to post-process completed jobs on Jono queues.
+
+mod harvester;
+
+pub use harvester::Harvester;
+
+pub mod prelude {
+    pub use crate::Harvester;
+}
